requirements.txt
altgraph==0.17.3
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
packaging==23.1
Pillow==9.5.0
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
PySide2==5.15.2.1
pytesseract==0.3.10
python-dotenv==1.0.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.30.0
shiboken2==5.15.2.1
tesseract==0.1.3
tk==0.1.0
urllib3==2.0.2
